+++ lib/neo4j/core/logging.rb
@@ -1,0 +1,40 @@
+# Copied largely from activerecord/lib/active_record/log_subscriber.rb
+module Neo4j
+  module Core
+    module Logging
+      class << self
+        def first_external_path_and_line(callstack)
+          line = callstack.find do |frame|
+            frame.absolute_path && !ignored_callstack(frame.absolute_path)
+          end
+
+          offending_line = line || callstack.first
+
+          [offending_line.path,
+           offending_line.lineno]
+        end
+
+        NEO4J_CORE_GEM_ROOT = File.expand_path('../../..', __dir__) + '/'
+
+        def ignored_callstack(path)
+          path.start_with?(NEO4J_CORE_GEM_ROOT) ||
+            path.start_with?(RbConfig::CONFIG['rubylibdir']) ||
+            path.start_with?(neo4j_gem_path) ||
+            path.start_with?(active_support_gem_path)
+        end
+
+        def neo4j_gem_path
+          return if !defined?(::Rails.root)
+
+          @neo4j_gem_path ||= File.expand_path('../../..', Neo4j::ActiveBase.method(:current_session).source_location[0])
+        end
+
+        def active_support_gem_path
+          return if !defined?(::ActiveSupport::Notifications)
+
+          @active_support_gem_path ||= File.expand_path('../../..', ActiveSupport::Notifications.method(:subscribe).source_location[0])
+        end
+      end
+    end
+  end
+end